Added ability to use a custom schema with custom field:value pairs in addition to payload.
If no custom schema is provided, it will default to creating a single field named 'payload' with a value of the payload string passed in. This however will not automatically create a PAYLOAD, just a searchable field called 'payload'. The index function still needs to be passed the 'PAYLOAD' option, followed by the payload, in order to store that payload. There are cases where we want the searchable fields' data to differ from the the actual payload stored.